# docs/employee/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Leave Management API",
    "description": "API documentation for managing users, leave types, and roles in the Leave Management System."
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
"/leave/view-all": {
          "get": {
            "summary": "View all leave requests",
            "operationId": "viewAllRequest",
            "tags": ["Leave Requests"],
            "responses": {
              "200": {
                "description": "Successfully retrieved leave requests",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "leaveRequests": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "userId": {
                                "type": "object",
                                "properties": {
                                  "firstName": { "type": "string" },
                                  "lastName": { "type": "string" },
                                  "email": { "type": "string" }
                                }
                              },
                              "description": { "type": "string" },
                              "startDate": { "type": "string", "format": "date" },
                              "endDate": { "type": "string", "format": "date" },
                              "status": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "error": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/leave/send-request": {
          "post": {
            "summary": "Submit a leave request",
            "operationId": "sendRequest",
            "tags": ["Leave Requests"],
            "requestBody": {
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "leaveTypeId": { "type": "string" },
                      "description": { "type": "string" },
                      "startDate": { "type": "string", "format": "date" },
                      "endDate": { "type": "string", "format": "date" }
                    },
                    "required": ["leaveTypeId", "description", "startDate", "endDate"]
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Successfully submitted leave request",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "Request sent successfully" },
                        "leave": {
                          "type": "object",
                          "properties": {
                            "userId": { "type": "string" },
                            "supervisorId": { "type": "string" },
                            "leaveTypeId": { "type": "string" },
                            "description": { "type": "string" },
                            "startDate": { "type": "string", "format": "date" },
                            "endDate": { "type": "string", "format": "date" },
                            "numberOfDays": { "type": "integer" }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "400": {
                "description": "Bad Request",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "Fill all fields in the form" }
                      }
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "error": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/leave/confirm/{id}": {
          "put": {
            "summary": "Approve a leave request",
            "operationId": "confirmRequest",
            "tags": ["Leave Requests"],
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Successfully approved leave request",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "Leave request Approved" },
                        "leaveRequest": {
                          "type": "object",
                          "properties": {
                            "userId": { "type": "string" },
                            "supervisorId": { "type": "string" },
                            "leaveTypeId": { "type": "string" },
                            "description": { "type": "string" },
                            "startDate": { "type": "string", "format": "date" },
                            "endDate": { "type": "string", "format": "date" },
                            "status": { "type": "string", "example": "Approved" }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "404": {
                "description": "Leave request not found",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "No request found" }
                      }
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "error": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "/leave/deny/{id}": {
          "put": {
            "summary": "Reject a leave request",
            "operationId": "denyRequest",
            "tags": ["Leave Requests"],
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Successfully rejected leave request",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "Leave request Rejected" },
                        "leaveRequest": {
                          "type": "object",
                          "properties": {
                            "userId": { "type": "string" },
                            "supervisorId": { "type": "string" },
                            "leaveTypeId": { "type": "string" },
                            "description": { "type": "string" },
                            "startDate": { "type": "string", "format": "date" },
                            "endDate": { "type": "string", "format": "date" },
                            "status": { "type": "string", "example": "Denied" }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "404": {
                "description": "Leave request not found",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "No request found" }
                      }
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "error": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
        "components": {
          "securitySchemes": {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        }
}